src/pybind/mgr/requirements-required.txt
-e ../../python-common
asyncmock
cherrypy
cryptography
jsonpatch
Jinja2
pecan
prettytable<3.4.0
pyfakefs
pyOpenSSL
pytest-cov==2.7.1
pyyaml
requests-mock
scipy
werkzeug
natsort
bcrypt
cheroot~=10.0
pyjwt
zstandard
